# Get rid of some `clone`s