Small optimisations:

1. A lot of filters consume `HashMap`s but unnecessarily clone the entries in them.
2. Filters which take array `Value`s do similar.
3. In `template.rs`, `Block`s and `Macro`s were cloned when a clone wasn't necessary.