@clarcharr 2017-01-14
Benchmarks on my CPU:

current master:
```
test bench_escape_html                         ... bench:         184 ns/iter (+/- 3)
test bench_parsing_basic_template              ... bench:      16,045 ns/iter (+/- 146)
test bench_parsing_with_inheritance_and_macros ... bench:      53,742 ns/iter (+/- 452)
test bench_rendering_basic_template            ... bench:      11,014 ns/iter (+/- 93)
test bench_rendering_inheritance_and_macros    ... bench:      23,595 ns/iter (+/- 179)
test bench_rendering_only_inheritance          ... bench:       5,884 ns/iter (+/- 47)
test bench_rendering_only_macro_call           ... bench:       3,865 ns/iter (+/- 44)
test bench_rendering_only_parent               ... bench:       3,652 ns/iter (+/- 51)
test bench_rendering_only_variable             ... bench:       3,156 ns/iter (+/- 21)
```

this PR:
```
test bench_escape_html                         ... bench:         168 ns/iter (+/- 4)
test bench_parsing_basic_template              ... bench:      15,562 ns/iter (+/- 186)
test bench_parsing_with_inheritance_and_macros ... bench:      51,638 ns/iter (+/- 454)
test bench_rendering_basic_template            ... bench:      10,929 ns/iter (+/- 74)
test bench_rendering_inheritance_and_macros    ... bench:      23,715 ns/iter (+/- 291)
test bench_rendering_only_inheritance          ... bench:       5,982 ns/iter (+/- 32)
test bench_rendering_only_macro_call           ... bench:       3,902 ns/iter (+/- 55)
test bench_rendering_only_parent               ... bench:       3,569 ns/iter (+/- 26)
test bench_rendering_only_variable             ... bench:       3,203 ns/iter (+/- 15)
```